shifted logic from 'load_from_api' into 'ingest_logic'
